requirements.txt
numpy
streamlit-drawable-canvas
streamlit~=1.52.2
pillow~=12.1.0
pandas~=2.3.3
matplotlib~=3.10.8
scikit-learn~=1.8.0
scipy~=1.16.3